The original, unchanged source from include-what-you-use.
This project is a fork of the fix_includes.py script from
   https://github.com/include-what-you-use/include-what-you-use
This commit just brings that in wholesale, without any changes yet.

It also includes the license file from iwyu -- we inherit that as
well.

I also updated the README to start to have a bunch of info about the
project.

Test Plan:
None