The beginning of the refactor of fix_includes.py.
My goal is to extract all the common functionality, shared by all
languages, into a base class or two in base.py.  Then each language
will have its own subclass.

In this commit, I exract most of the functionality for the base class,
but not all of it.  I also haven't even tried running it -- this is
just a checkpointing commit.