Fix a few bugs in found by using the code.
1) The docstring-detection RE was failing due to a missing DOTALL.
2) The is-system-include check was failing on nested modules
   such as xml.etree.  I wasn't chopping off at the first dot when
   checking whether a module was in the system-list or not.
3) My "does this include-range overlap with contentful content"
   was wrong, because I was treating the end of the range as
   inclusive when it's not: that is, I was treating the range
   as `[a, b]` when it's really `[a, b)`.  Thus, I'm ok as long
   as first_contentful_line >= b, but I was doing a > check instead.

Test Plan:
I ran
   echo 'user_models.py should add these lines:\nimport deferred_util\n\n' | fix_python_imports.py -n
and saw it did not put deferred_util at the top of the file (this is
due to fix #3).

I also saw, before fixing #3, that it put deferred_util.py after the
top-of-file docstring rather than on line 1 (this is due to fix #1).

Then I ran
   fix_python_imports.py -n -s api/internal/sat.py
and saw that it kept `import xml.etree.ElementTree` with the system
imports.

I've put this data into a test-file TODO but haven't yet made python
tests.