Improve categorization of system vs. pip-installed modules
Summary:
Previously, we used `pkgutil.iter_modules()`, which basically just looks at
everything on `sys.path` to get the "system" modules.  This wasn't great, since
it includes everything from the virtualenv, and it was even worse if you called
the script from another script which might, for example, have all of webapp on
its path -- now everything will look like a system module.

Now, instead, we split up the path ourselves, into what we think is local vs.
third-party vs. system.  It's unfortunately a bit of guesswork; python doesn't
really know the difference.  But we can guess well enough to get it mostly
right.

This allows removing most of our special-cases -- we just need to keep the ones
that are physically local even though we consider them third-party (i.e.
`third_party` and `shared`), and the frankenserver imports, since those get
added to the path later (and thus may not be on it in this script).

Test Plan:
- Run `fix_python_imports.py -s` on some files, see it sort imports correctly.
- Do the same from outside of the virtualenv, and from a script with webapp on
  `sys.path`.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D37810