Change how fix_python_imports categorizes imports, again
Summary:
This is a follow-on -- honestly a replacement for -- D37810 which changed how
we categorize system and third-party imports.  The particular issue that it
fixes is that we weren't correctly categorizing things that live in
Frankenserver that aren't `google.*`, i.e., things from the `lib` directory.
More generally we are now a bit smarter about looking at where you would import
a module from rather than just its name -- so this will for example work right
if we don't do `third_party.*` for third party imports.

What we do now is basically we just look at the modules on your path, and where
they would be imported from, and then guess based on *that* directory.  We
assume that anything not relative to cwd that contains "site-packages" or
"dist-packages" must be a third-party package directory, and that anything
relative to cwd that contains "shared" or "third_party" must be some sort of
vendored thing.  This allows us to remove all of the special cases by module
name.  It does make us more dependent on having an approximately correct
PYTHONPATH, but that's fairly easy for callers to do.

Well, for most callers.  There is one annoying KA-specific hack, which is that
we try running `fix_sys_path()`, if we can, to make sure our path looks like
webapp's, before running imports.  (If we don't find it, we just continue, so
this won't prevent any non-webapp uses.)  Ah well.

Test Plan:
In webapp, run `PYTHONPATH=. path/to/fix_python_imports.py -s content/*.py`;
look over the changes manually to confirm that `google.*`, `shared.*`,
`third_party.*`, `mock` (from virtualenv), `datetime` (builtin), `logging`
(python system module), and `webapp2` (from frankenserver lib) are all sorted
in their correct places.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D38309