Allow docstrings to occur other than at the very beginning of a file.
Sometimes docstrings occur later, e.g. if there's a shebang line at
the top of the file, or a comment indicating the encoding.

We still only accept docstrings that start at the very beginning of
the line, so these are still top-level docstrings, not per-function
docstrings or the like (which we don't care about).

Auditors: benkraft

Test plan:
I ran
   ( echo "api/auth/decorators.py should add these lines:" ; echo "import auth.tokens" ; echo ) | env PYTHONPATH=.:$PYTHONPATH ~/khan/fix-includes/fix-includes/fix_python_imports.py
where api/auth/decorators.py starts like this:
```
# TODO(colin): fix these lint errors (http://pep8.readthedocs.io/en/release-1.7.x/intro.html#error-codes)
# pep8-disable:E265
"""Decorators for verifying authentication and authorization of users.

...
"""
from google.appengine.ext import db
...
```

Before this change, it added the new import at the top of the file,
afterwards it added it after the `from` line like it should.