Fix a really silly mistake.
This is the first time I've ever tried to use this script to delete
import lines, which is why we hadn't caught it before.  I don't know
what I was thinking when I typed it.

Auditors: benkraft

Test plan:
Ran a fix-python-imports run with some removal-text and had it run
without error.