Add a `--root` flag to specify project-root, rather than assuming os.getcwd().
Summary:
This lets you run fix-python-imports from a different directory than
the rootdir of your project.

Test Plan:
I ran
   env PYTHONPATH=. ~/khan/fix-includes/fix-includes/fix_python_imports.py -s request_handler.py
and it didn't suggest any changes, which indicates it doesn't break
existing usage at least.

I'd like to get rid of the need for `PYTHONPATH=.` -- we'd do this by
always looking at the importers for the `--root` dir -- but I don't
fully understand the repercussions of that so I don't do it yet.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D40324