Cache _SYSTEM_MODULES/_THIRD_PARTY_MODULES by rootdir.
Every time the rootdir changes, the definition of what's a third-party
module changes as well.  When run from the commandline it won't change
within the run of a file, but it could when used as a library!  This
makes it safer, at some cost in readability due to having to pass the
info around everywhere.

Auditors: benkraft

Test Plan:
I tried using this in slicker and it gave the right results.