Change console script names to avoid package vs. script collisions
Summary:
Turns out `virtualenv` sets things up such that its `bin` is on the
default `sys.path`.  This meant that if another script tried to do
`import fix_includes`, it got the console script, not the package.
That's no good!  To fix it, I renamed the console scripts to be
extensionless-with-dashes, which is a better naming convention anyway.
Setuptools makes this easy via `entry_points`.  The only annoying part
is it wants to call a function, whereas we previously had slightly
nontrivial `if __name__ == '__main__':` blocks, so I wrapped those up in
a function.

Test Plan:
- `python setup.py install`
- `fix-python-imports -s -n main.py` works in webapp
- `import fix_includes.fix_python_imports` works in a python shell

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D42145